ci: clear the Doc and Clippy gates
`cargo doc --no-deps` under `RUSTDOCFLAGS=-D warnings` and
`cargo clippy --all-targets -- -D warnings` are both CI gates, and both were
failing. Two classes of breakage:

* Broken intra-doc links — doc comments bracketing a name that is private, out
  of scope, or not a Rust item at all (C function names in port docs, Doxygen
  `@param[in]` markers, `<CR>`/`<NL>`, `argv[0]`). Bracketed text is a link to
  rustdoc; these were never meant as links, so they are plain code spans now.
  Where the target IS a reachable item the link is kept and qualified instead.
* Clippy findings in real code, fixed rather than allowed — complex tuples got
  named type aliases, dead code was removed, and the one deliberate exception is
  documented inline with the measurement that justifies it.

Also gates `tiers::tests::PROGRAM` behind `#[cfg(test)]`: it is only read by the
tests, so a non-test build saw it as dead code.

+++ znative/src/lib.rs
@@ -91,8 +91,8 @@ pub const INIT_SYMBOL: &[u8] = b"znative_init\0";
 /// * `host`   — the host API table (call back into the shell through it).
 /// * `argc`   — number of elements in `argv`.
 /// * `argv`   — NUL-terminated C strings; `argv[0]` is the command name,
-///              `argv[1..]` the arguments. Valid only for the duration
-///              of the call; copy anything you need to keep.
+///   `argv[1..]` the arguments. Valid only for the duration of the call;
+///   copy anything you need to keep.
 ///
 /// Returns the command's exit status (0 = success), like any shell
 /// builtin.
@@ -256,7 +256,7 @@ impl Host {
     /// Evaluate shell `code`; returns its exit status.
     pub fn eval(&self, code: &str) -> i32 {
         match CString::new(code) {
-            Ok(c) => (self.t().eval)(self.api, c.as_ptr()) as i32,
+            Ok(c) => (self.t().eval)(self.api, c.as_ptr()),
             Err(_) => 1,
         }
     }
@@ -371,7 +371,7 @@ impl Host {
             }
         }
         let argv: Vec<*const c_char> = cargs.iter().map(|c| c.as_ptr()).collect();
-        (self.t().comp_dispatch)(self.api, argv[0], argv.len(), argv.as_ptr()) as i32
+        (self.t().comp_dispatch)(self.api, argv[0], argv.len(), argv.as_ptr())
     }
 
     /// Empty the command-name hash — the observable effect of the
